refactor([issue-743]): split identity.js into focused submodules
Break the 1917-LOC server/services/identity.js into a barrel plus eight
submodules under server/services/identity/ (store, markers, chronotype,
longevity, goals, todos, insights, status). The barrel re-exports the exact
same 43 identifiers, so every importer keeps working unchanged. Pure
structural move — no behavior change.
